# pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/request_reset_token/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Request Reset Token",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httprequest.AuthenticationRequestResetToken"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/auth/validate_reset_token/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Validate Token Reset Password",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/httprequest.AuthenticationValidateResetToken"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "httprequest.AuthenticationRequestResetToken": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "httprequest.AuthenticationValidateResetToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer access token here",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}